requirements.txt
transformers===4.27.2
tensorflow-text===2.7.3
websockets===10.4
rasa===3.2.4
rasa-sdk===3.2.0
sanic===21.12.2
python-dotenv===0.20.0
pandas===1.4.3
requests===2.28.1
fbmessenger===6.0.0
sentence-transformers==2.2.2
openpyxl==3.0.10

